+++ src/OTOWNDRW.cpp
@@ -263,7 +263,8 @@ int Town::draw_detect_link_line(int actionDetect)
 
 			else if( rc==2
 				&& (!Ambition::Config::enhancementsAvailable()
-					|| !(mouse.event_skey_state & ALT_KEY_MASK)))
+					|| (!(mouse.event_skey_state & ALT_KEY_MASK)
+						&& !(mouse.event_skey_state & SHIFT_KEY_MASK))))
 			{
 				if( firmPtr->nation_recno == nation_recno && !firmPtr->under_construction &&
 					 firmPtr->worker_array && firmPtr->worker_count < MAX_WORKER )