+++ src/ambition/version_details.cc
@@ -7,7 +7,7 @@ namespace Ambition::News {
 
 std::vector<VersionUpdates> versionDetails = {
   {
-    .identifier = _("Preview Version -- not yet released"),
+    .identifier = _("3.1.3 -- 2025-04-27"),
     .sections = {
       {
         .heading = _("Added"),